backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mistralai>=1.0.0
gitpython==3.1.43
python-dotenv==1.0.1
pydantic-settings==2.5.2
aiofiles==24.1.0
websockets==12.0
